+++ opbeat/contrib/django/models.py
@@ -16,6 +16,7 @@ from __future__ import absolute_import
 import sys
 import logging
 import warnings
+from opbeat.contrib.django.utils import disabled_due_to_debug
 from opbeat.utils import six
 
 from django.conf import settings as django_settings
@@ -131,7 +132,8 @@ def get_client(client=None):
             organization_id=config.get('ORGANIZATION_ID', None),
             app_id=config.get('APP_ID', None),
             secret_token=config.get('SECRET_TOKEN', None),
-            processors=config.get('PROCESSORS', None)
+            processors=config.get('PROCESSORS', None),
+            async=config.get('ASYNC', None),
         )
         if not tmp_client:
             _client = (client, instance)
@@ -139,40 +141,16 @@ def get_client(client=None):
     return _client[1]
 
 
-def get_transaction_wrapper(client):
-    if client.servers:
-        class MockTransaction(object):
-            def commit_on_success(self, func):
-                return func
-
-            def is_dirty(self):
-                return False
-
-            def rollback(self):
-                pass
-
-        transaction = MockTransaction()
-    else:
-        from django.db import transaction
-
-    return transaction
-
-
 def opbeat_exception_handler(request=None, **kwargs):
-    transaction = get_transaction_wrapper(client)
-
-    @transaction.commit_on_success
     def actually_do_stuff(request=None, **kwargs):
         exc_info = sys.exc_info()
         try:
-            config = getattr(django_settings, 'OPBEAT', {})
-            if (django_settings.DEBUG and not config.get('DEBUG', False)) or getattr(exc_info[1], 'skip_opbeat', False):
+            if (disabled_due_to_debug()
+                    or getattr(exc_info[1], 'skip_opbeat', False)):
                 return
 
-            if transaction.is_dirty():
-                transaction.rollback()
-
-            get_client().capture('Exception', exc_info=exc_info, request=request)
+            get_client().capture('Exception', exc_info=exc_info,
+                                 request=request)
         except Exception as exc:
             try:
                 logger.exception(u'Unable to process log entry: %s' % (exc,))
@@ -185,6 +163,7 @@ def opbeat_exception_handler(request=None, **kwargs):
                 logger.exception(e)
 
     return actually_do_stuff(request, **kwargs)
+
 
 def register_handlers():
     from django.core.signals import got_request_exception