+++ opbeat/utils/compat.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+import atexit
 import functools
 
 try:
@@ -14,3 +15,25 @@ def noop_decorator(func):
     def wrapped(*args, **kwargs):
         return func(*args, **kwargs)
     return wrapped
+
+
+def atexit_register(func):
+    """
+    Uses either uwsgi's atexit mechanism, or atexit from the stdlib.
+
+    When running under uwsgi, using their atexit handler is more reliable,
+    especially when using gevent
+    :param func: the function to call at exit
+    """
+    try:
+        import uwsgi
+        orig = getattr(uwsgi, 'atexit', None)
+
+        def uwsgi_atexit():
+            if callable(orig):
+                orig()
+            func()
+
+        uwsgi.atexit = uwsgi_atexit
+    except ImportError:
+        atexit.register(func)