+++ tests/client/client_tests.py
@@ -8,7 +8,7 @@ import string
 from opbeat.utils import six
 from socket import socket, AF_INET, SOCK_DGRAM
 from opbeat.utils.compat import TestCase
-from opbeat.base import Client, ClientState
+from opbeat.base import Client, ClientState, DummyClient
 from opbeat.utils.stacks import iter_stack_frames
 
 from tests.helpers import get_tempstoreclient
@@ -108,7 +108,7 @@ class ClientTest(TestCase):
             headers={
                 'Content-Type': 'application/octet-stream',
                 'Authorization': 'Bearer %s' % (access_token),
-                'User-Agent': 'opbeat/%s' % opbeat.VERSION
+                'User-Agent': 'opbeat-python/%s' % opbeat.VERSION
             },
         )
 
@@ -154,7 +154,7 @@ class ClientTest(TestCase):
             headers={
                 'Content-Type': 'application/octet-stream',
                 'Authorization': 'foo',
-                'User-Agent': 'opbeat/%s' % opbeat.VERSION
+                'User-Agent': 'opbeat-python/%s' % opbeat.VERSION
 
             },
         )
@@ -318,28 +318,24 @@ class ClientTest(TestCase):
         event = self.client.events.pop(0)
         self.assertTrue(len(event['logger']) < 61, len(event['logger']))
 
-    # def test_long_server_name(self):
-    #     message = 's' * 201
+    @mock.patch('opbeat.base.Client.send')
+    @mock.patch('opbeat.base.RequestsStore.should_collect')
+    def test_metrics_collection(self, should_collect, mock_send):
+        client = Client(
+            servers=['http://example.com'],
+            organization_id='organization_id',
+            app_id='app_id',
+            secret_token='secret',
+        )
+        should_collect.return_value = False
+        for i in range(7):
+            client.captureRequest(0.1, 200, 'test')
 
-    #     self.client.capture('Message', message=message,)
+        self.assertEqual(len(client._requests_store), 7)
+        self.assertEqual(mock_send.call_count, 0)
+        should_collect.return_value = True
 
-    #     self.assertEquals(len(self.client.events), 1)
-    #     event = self.client.events.pop(0)
-    #     self.assertTrue(len(event['server_name']) < 201)
+        client.captureRequest(0.1, 200, 'test')
+        self.assertEqual(len(client._requests_store), 0)
+        self.assertEqual(mock_send.call_count, 1)
 
-# class ClientUDPTest(TestCase):
-#     def setUp(self):
-#         self.server_socket = socket(AF_INET, SOCK_DGRAM)
-#         self.server_socket.bind(('127.0.0.1', 0))
-#         self.client = Client(servers=["udp://%s:%s" % self.server_socket.getsockname()], key='BassOmatic')
-
-#     def test_delivery(self):
-#         self.client.create_from_text('test')
-#         data, address = self.server_socket.recvfrom(2**16)
-#         self.assertTrue("\n\n" in data)
-#         header, payload = data.split("\n\n")
-#         for substring in ("sentry_timestamp=", "sentry_client="):
-#             self.assertTrue(substring in header)
-
-#     def tearDown(self):
-#         self.server_socket.close()