+++ tests/contrib/django/django_tests.py
@@ -284,7 +284,7 @@ class DjangoClientTest(TestCase):
         self.opbeat.exclude_paths = ['tests.views.decorated_raise_exc']
         self.assertRaises(Exception, self.client.get, reverse('opbeat-raise-exc-decor'))
 
-        self.assertEquals(len(self.opbeat.events), 1)
+        self.assertEquals(len(self.opbeat.events), 1, self.opbeat.events)
         event = self.opbeat.events.pop(0)
 
         self.assertEquals(event['culprit'], 'tests.contrib.django.views.raise_exc')
@@ -447,6 +447,23 @@ class DjangoClientTest(TestCase):
     #         self.assertEquals(len(self.opbeat.events), 1)
     #         self.opbeat.events.pop(0)
 
+    def test_request_metrics(self):
+        self.opbeat._requests_store.get_all()  # clear the store
+        with self.settings(MIDDLEWARE_CLASSES=['opbeat.contrib.django.middleware.OpbeatAPMMiddleware']):
+            self.assertEqual(len(self.opbeat._requests_store), 0)
+            self.client.get(reverse('opbeat-no-error'))
+            self.assertEqual(len(self.opbeat._requests_store), 1)
+            timed_requests = self.opbeat._requests_store.get_all()
+
+            self.assertEqual(len(timed_requests), 1)
+            timing = timed_requests[0]
+            self.assertTrue('durations' in timing)
+            self.assertEqual(len(timing['durations']), 1)
+            self.assertEqual(timing['transaction'],
+                             'tests.contrib.django.views.no_error')
+            self.assertEqual(timing['result'],
+                             200)
+
 
 class DjangoLoggingTest(TestCase):
     def setUp(self):
@@ -524,12 +541,6 @@ class CeleryIntegratedClientTest(TestCase):
             app_id='app',
             secret_token='secret',
         )
-
-    @mock.patch('opbeat.contrib.django.celery.CeleryClient.send_raw_integrated')
-    def test_send_encoded(self, send_raw):
-        self.client.send_integrated('foo')
-
-        send_raw.delay.assert_called_once_with('foo')
 
     @skipIf(not has_with_eager_tasks, 'with_eager_tasks is not available')
     @with_eager_tasks